requirements.txt
Pyrogram
TgCrypto
python-dotenv
Flask

